+++ src/main/java/com/journey/journeybookshelf/controller/AuthenticationController.java
@@ -1,18 +1,19 @@
 package com.journey.journeybookshelf.controller;
 
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
+@Controller
 public class AuthenticationController {
 
-    // Login
-    //
     @GetMapping("/login")
     public String loginForm(){
         return "users/login";
     }
 
     @GetMapping("/error")
-    public String errorForm() { return "users/login";}
-
+    public String errorForm(){
+        return "users/login";
+    }
 
 }
