--- src/main/java/com/journey/journeybookshelf/models/UserwithRoles.java
@@ -1,40 +1,0 @@
-package com.journey.journeybookshelf.models;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.List;
-
-public class UserwithRoles extends User implements UserDetails {
-
-        public UserwithRoles(User user) {
-            super(user);  // Call the copy constructor defined in User
-        }
-
-        public List<GrantedAuthority> getAuthorities() {
-            String roles = ""; // Since we're not using the authorization part of the component
-            return AuthorityUtils.commaSeparatedStringToAuthorityList(roles);
-        }
-
-        @Override
-        public boolean isAccountNonExpired() {
-            return true;
-        }
-
-        @Override
-        public boolean isAccountNonLocked() {
-            return true;
-        }
-
-        @Override
-        public boolean isCredentialsNonExpired() {
-            return true;
-        }
-
-        @Override
-        public boolean isEnabled() {
-            return true;
-        }
-    }
-
