+++ src/main/java/com/journey/journeybookshelf/services/UserDetailsLoader.java
@@ -1,7 +1,7 @@
 package com.journey.journeybookshelf.services;
 
 import com.journey.journeybookshelf.models.User;
-import com.journey.journeybookshelf.models.UserwithRoles;
+import com.journey.journeybookshelf.models.UserWithRoles;
 import com.journey.journeybookshelf.repository.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -10,11 +10,10 @@ import org.springframework.stereotype.Service;
 
 
 @Service
-public class userDetailsLoader implements UserDetailsService {
+public class UserDetailsLoader implements UserDetailsService {
     private final UserRepository users;
 
-
-    public userDetailsLoader(UserRepository users) {
+    public UserDetailsLoader(UserRepository users) {
         this.users = users;
     }
 
@@ -25,7 +24,7 @@ public class userDetailsLoader implements UserDetailsService {
             throw new UsernameNotFoundException("No user found for " + username);
         }
 
-        return new UserwithRoles(user);
+        return new UserWithRoles(user);
     }
 }
