+++ src/test/java/com/journey/journeybookshelf/TheJourney2ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.journey.journeybookshelf;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TheJourney2ApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
